Extract "declare" tests into separate test class
