+++ drools-compiler/src/test/resources/org/drools/compiler/integrationtests/drl/test_DeclareWithFrom.drl
